pyproject.toml
[tool.poetry]
name = "auth-service"
version = "0.1.0"
description = "Serviço de autenticação para Tech Challenge"
authors = ["magnifficco.io"]
readme = "README.md"
license = "MIT"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.8"
sqlalchemy = "^2.0.38"
alembic = "^1.14.1"
uvicorn = "^0.34.0"
pydantic = "^2.10.6"
bcrypt = "^4.2.1"
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
python-multipart = "^0.0.20"



[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.25.3"
black = "^25.1.0"
isort = "^6.0.0"
mypy = "^1.15.0"
pytest-mock = "^3.14.0"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "app.main:run_server"